+++ iosApp/iosApp/Model/Data/AgendaContent.swift
@@ -9,6 +9,7 @@
 import Foundation
 import shared
 
+// MARK: - AgendaContent
 struct AgendaContent {
     struct Session: Hashable, Identifiable {
         let id: String
@@ -29,6 +30,7 @@ struct AgendaContent {
         var isATalk: Bool { return (openFeedbackFormId != nil) }
     }
     
+// MARK: - Section
     struct Section: Hashable {
         let date: Date
         let day: String
@@ -38,8 +40,8 @@ struct AgendaContent {
     var sections: [Section]
 }
 
- extension AgendaContent.Session {
-    
+extension AgendaContent.Session {
+    //Initialization session
     init(from session: Session) {
         let newFormatter = ISO8601DateFormatter()
         self.init(id: session.id, abstract: session.abstract, category: session.category, language: session.language, complexity: session.complexity, openFeedbackFormId: session.openFeedbackFormId, speakers: session.speakers, room: session.room?.name ?? "", date: newFormatter.date(from: session.scheduleSlot.startDate) ?? Date(), startDate: session.scheduleSlot.startDate, endDate: session.scheduleSlot.endDate,durationAndLanguage: session.getDurationAndLanguageString(), title: session.title, sessionType: session.type)
