+++ iosApp/iosApp/iOSApp.swift
@@ -8,19 +8,19 @@ import FirebaseCore
 import FirebaseCrashlytics
 import FirebaseRemoteConfig
 
+///RemoteConfig properties
 var remoteConfig = RemoteConfig.remoteConfig()
 
 @main
 struct iOSApp: App {
     
     init() {
-        //NSLogger
+        //NSLogger initialization
         LoggerSetupBonjourForBuildUser()
         
-        //Firebase
+        //Firebase initialization
         FirebaseApp.configure()
         _ = RCValues.sharedInstance
-        
     }
     
     var body: some Scene {
